+++ include/xgboost/feature_map.h
@@ -79,7 +79,7 @@ class FeatureMap {
     if (!strcmp("i", tname)) return kIndicator;
     if (!strcmp("q", tname)) return kQuantitive;
     if (!strcmp("int", tname)) return kInteger;
-    if (!strcmp("float", tname)) return kFloat;
+    if (!strcmp("double", tname)) return kFloat;
     LOG(FATAL) << "unknown feature type, use i for indicator and q for quantity";
     return kIndicator;
   }
